Convert to Maven project.
Added a pom.xml and added those dependencies that were required to solve all compile errors. Basically, GNormPlus required the BioC API and the snowball stemmer. Unfortunately, GNormPlus uses some old version of the BioC API that is not available on Maven I think. Thus I had to adapt the BioCDoc class to the BioC API that is available on Maven. No large changes but quite a few of them.
I used the 'jar-with-dependencies' Maven artifact (built with the `mvn package` command) to re-run some evaluations and the numbers match exactly with the original GNormPlus.