Fix bugs and check performance.
The BioCDoc class had a bug introduces with the conversion script because of the optionals for annotation infon values. Those have been fixed manually in this release.
Tested the performance in a pipeline and it was virtually equal to the stand-alone system. This version had 3FPs less than the original for some reason on NLM-Gene test but that's alright for me ;-)