Fix #1.
The issue was that in SR an intermediate BioC file is read in. Annotations were reset before, PMIDs were not and then the new data was just read into the old BioCDocobj, leading to a duplication of the PMID list. The GeneIDMatch method iterates over the PMIDs and thus misses annotations for the second half of PMIDs - which are duplicates anyway.